+++ src/routes/AccountsListingContainer.js
@@ -72,7 +72,7 @@ class AccountsListingContainer extends React.Component {
       records: 'feefineactions',
       path: 'feefineactions?query=(userId==:{id} and comments=*)&limit=%{activeRecord.comments}',
       shouldRefresh: (resource, action, refresh) => {
-        return refresh || action.meta.path === 'accounts';
+        return refresh || action.meta.path === 'accounts' || action.meta.path === 'accounts-bulk';
       },
     },
     filter: {